karyon_jsonrpc::client

Struct Client

pub struct Client<C> {
    disconnect: AtomicBool,
    message_dispatcher: MessageDispatcher,
    subscriptions: Arc<Subscriptions>,
    send_chan: (Sender<Value>, Receiver<Value>),
    task_group: TaskGroup,
    config: ClientConfig,
    codec: C,
}

Represents an RPC client

Fields
disconnect: AtomicBoolmessage_dispatcher: MessageDispatchersubscriptions: Arc<Subscriptions>send_chan: (Sender<Value>, Receiver<Value>)task_group: TaskGroupconfig: ClientConfigcodec: C
Implementations

impl<C> Client<C>
where C: ClonableJsonCodec + 'static,

pub async fn call<T: Serialize + DeserializeOwned, V: DeserializeOwned>( &self, method: &str, params: T, ) -> Result<V>

Calls the provided method, waits for the response, and returns the result.

pub async fn subscribe<T: Serialize + DeserializeOwned>( &self, method: &str, params: T, ) -> Result<Arc<Subscription>>

Subscribes to the provided method, waits for the response, and returns the result.

This function sends a subscription request to the specified method with the given parameters. It waits for the response and returns a Subscription.

pub async fn unsubscribe( &self, method: &str, sub_id: SubscriptionID, ) -> Result<()>

Unsubscribes from the provided method, waits for the response, and returns the result.

This function sends an unsubscription request for the specified method and subscription ID. It waits for the response to confirm the unsubscription.

pub async fn stop(&self)

Disconnect the client

async fn send_request<T: Serialize + DeserializeOwned>( &self, method: &str, params: T, ) -> Result<Response>

async fn send(&self, req: Request) -> Result<()>

async fn init(config: ClientConfig, codec: C) -> Result<Arc<Self>>

Initializes a new Client from the provided ClientConfig.

async fn connect(self: &Arc<Self>) -> Result<Conn<Value, Error>>

fn start_background_loop(self: &Arc<Self>, conn: Conn<Value, Error>)

async fn background_loop( self: Arc<Self>, conn: Conn<Value, Error>, ) -> Result<()>

async fn handle_msg(&self, msg: Value) -> Result<()>