karyon_jsonrpc::server::builder
Struct ServerBuilder

pub struct ServerBuilder<C> {
    inner: ServerConfig<C>,
}

Builder for constructing an RPC Server.

Fields
inner: ServerConfig<C>
Implementations

impl<C> ServerBuilder<C>
where C: ClonableJsonCodec,

pub fn service(self, service: Arc<dyn RPCService>) -> Self

Adds a new RPC service to the server.

Example

use std::sync::Arc;

use serde_json::Value;

use karyon_jsonrpc::{Server, rpc_impl, RPCError};

struct Ping {}

#[rpc_impl]
impl Ping {
    async fn ping(&self, _params: Value) -> Result<Value, RPCError> {
        Ok(serde_json::json!("Pong"))
    }
}

async {
    let server = Server::builder("ws://127.0.0.1:3000")
        .expect("Create a new server builder")
        .service(Arc::new(Ping{}))
        .build().await
        .expect("Build the server");
};

pub fn pubsub_service(self, service: Arc<dyn PubSubRPCService>) -> Self

Adds a new PubSub RPC service to the server.

Example

use std::sync::Arc;

use serde_json::Value;

use karyon_jsonrpc::{
    Server, rpc_impl, rpc_pubsub_impl, RPCError, Channel, SubscriptionID,
};

struct Ping {}

#[rpc_impl]
impl Ping {
    async fn ping(&self, _params: Value) -> Result<Value, RPCError> {
        Ok(serde_json::json!("Pong"))
    }
}

#[rpc_pubsub_impl]
impl Ping {
   async fn log_subscribe(
        &self,
        chan: Arc<Channel>,
        method: String,
        _params: Value,
    ) -> Result<Value, RPCError> {
        let sub = chan.new_subscription(&method).await;
        let sub_id = sub.id.clone();
        Ok(serde_json::json!(sub_id))
    }

    async fn log_unsubscribe(
        &self,
        chan: Arc<Channel>,
        _method: String,
        params: Value,
    ) -> Result<Value, RPCError> {
        let sub_id: SubscriptionID = serde_json::from_value(params)?;
        chan.remove_subscription(&sub_id).await;
        Ok(serde_json::json!(true))
    }
}

async {
    let ping_service = Arc::new(Ping{});
    let server = Server::builder("ws://127.0.0.1:3000")
        .expect("Create a new server builder")
        .service(ping_service.clone())
        .pubsub_service(ping_service)
        .build().await
        .expect("Build the server");
};

pub fn tcp_config(self, config: TcpConfig) -> Result<ServerBuilder<C>>

Configure TCP settings for the server.

Example

use karyon_jsonrpc::{Server, TcpConfig};

async {
    let tcp_config = TcpConfig::default();
    let server = Server::builder("ws://127.0.0.1:3000")
        .expect("Create a new server builder")
        .tcp_config(tcp_config)
        .expect("Add tcp config")
        .build().await
        .expect("Build the server");
};

This function will return an error if the endpoint does not support TCP protocols.

pub async fn build(self) -> Result<Arc<Server<C>>>

Builds the server with the configured options.

pub async fn build_with_executor(self, ex: Executor) -> Result<Arc<Server<C>>>

Builds the server with the configured options and an executor.