karyon_jsonrpc::server

Struct Server

pub struct Server {
    listener: Listener<Value, Error>,
    task_group: TaskGroup,
    config: ServerConfig,
}

Represents an RPC server

Fields
listener: Listener<Value, Error>task_group: TaskGroupconfig: ServerConfig
Implementations

impl Server

pub fn start(self: Arc<Self>)

Starts the RPC server by spawning a new task for the main accept loop. The accept loop listens for incoming connections.

This function does not block the current thread. If you need the thread to block, use the [start_block] method instead.

pub async fn start_block(self: Arc<Self>) -> Result<()>

Starts the RPC server by running the main accept loop. The accept loop listens for incoming connections and blocks the current thread.

If you prefer a non-blocking implementation, use the [start] method instead.

async fn accept_loop(self: &Arc<Self>) -> Result<()>

pub fn local_endpoint(&self) -> Result<Endpoint>

Returns the local endpoint.

pub async fn shutdown(&self)

Shuts down the RPC server

async fn handle_conn(self: &Arc<Self>, conn: Conn<Value, Error>) -> Result<()>

Handles a new connection

fn sanity_check(&self, request: Value) -> SanityCheckResult

async fn new_request( self: &Arc<Self>, queue: Arc<ResponseQueue<Value>>, channel: Arc<Channel>, msg: Value, )

Spawns a new task for handling the new request

async fn handle_request(&self, channel: Arc<Channel>, msg: Value) -> Response

Handles the new request, and returns an RPC Response that has either an error or result

async fn init( config: ServerConfig, ex: Option<Executor>, codec: impl ClonableJsonCodec + 'static, ) -> Result<Arc<Self>>

Initializes a new Server from the provided ServerConfig

async fn listen( config: &ServerConfig, codec: impl ClonableJsonCodec + 'static, ) -> Result<Listener<Value, Error>>