karyon_jsonrpc

Module message

Structs

Error

Notification

Notification represents a JSON-RPC notification message.

NotificationResult

NotificationResult represents the result of a subscription notification. It includes the result and the subscription ID that triggered the notification.

Request

Request represents a JSON-RPC request message. It includes the JSON-RPC version, an identifier for the request, the method to be invoked, and optional parameters.

Response

Response represents a JSON-RPC response message. It includes the JSON-RPC version, an identifier matching the request, the result of the request, and an optional error.

Constants

INTERNAL_ERROR_CODE

Internal error: Internal JSON-RPC error.

INTERNAL_ERROR_MSG

INVALID_PARAMS_ERROR_CODE

Invalid params: Invalid method parameter(s).

INVALID_REQUEST_ERROR_CODE

Invalid request: The JSON sent is not a valid Request object.

JSONRPC_VERSION

METHOD_NOT_FOUND_ERROR_CODE

Method not found: The method does not exist / is not available.

PARSE_ERROR_CODE

Parse error: Invalid JSON was received by the server.

Type Aliases

ID

SubscriptionID

SubscriptionID is used to identify a subscription.